feat(pages, constants, components): Set up starter pages and file structure
Created initial pages directories for the project: login, register, messages, profile, and homepage.
Each route folder includes a `page.tsx` file and a corresponding `module.css` for styling.
Additionally, a `constants/routes.ts` file was added to store the current routes for the project.

In the app/ directory, a `components` folder was created to house the components for the homepage.

Also, unnecessary files in the fonts and public folders were deleted to clean up the project.